Refactor RenderGraph into its own namespace
