+++ src/graphics/graphics/vulkan/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(core)
+add_subdirectory(render_graph)
 
 target_sources(${PROJECT_GRAPHICS_VULKAN_HEADLESS_LIB} PRIVATE
     GPUSceneContext.cc
