+++ src/graphics/graphics/vulkan/GuiRenderer.cc
@@ -3,11 +3,11 @@
 #include "assets/Asset.hh"
 #include "assets/AssetManager.hh"
 #include "graphics/gui/GuiManager.hh"
+#include "graphics/vulkan/VertexLayouts.hh"
 #include "graphics/vulkan/core/CommandContext.hh"
 #include "graphics/vulkan/core/Common.hh"
 #include "graphics/vulkan/core/DeviceContext.hh"
 #include "graphics/vulkan/core/Util.hh"
-#include "graphics/vulkan/core/Vertex.hh"
 
 #include <GLFW/glfw3.h>
 #include <algorithm>
