+++ src/graphics/graphics/vulkan/Renderer.hh
@@ -11,7 +11,7 @@
 #include "graphics/vulkan/core/Common.hh"
 #include "graphics/vulkan/core/Memory.hh"
 #include "graphics/vulkan/core/PerfTimer.hh"
-#include "graphics/vulkan/core/RenderGraph.hh"
+#include "graphics/vulkan/render_graph/RenderGraph.hh"
 
 #include <atomic>
 #include <functional>
@@ -30,8 +30,8 @@ namespace sp {
 namespace sp::vulkan {
     class Model;
     class GuiRenderer;
-    class RenderGraph;
-    class RenderGraphResources;
+
+    namespace rg = render_graph;
 
     struct LightingContext {
         int count = 0;
@@ -74,13 +74,13 @@ namespace sp::vulkan {
         void EndFrame();
 
         struct DrawBufferIDs {
-            RenderGraphResourceID drawCommandsBuffer; // first 4 bytes are the number of draws
-            RenderGraphResourceID drawParamsBuffer;
+            rg::ResourceID drawCommandsBuffer; // first 4 bytes are the number of draws
+            rg::ResourceID drawParamsBuffer;
         };
 
         DrawBufferIDs GenerateDrawsForView(ecs::Renderable::VisibilityMask viewMask);
         void DrawSceneIndirect(CommandContext &cmd,
-            RenderGraphResources &resources,
+            rg::Resources &resources,
             BufferPtr drawCommandsBuffer,
             BufferPtr drawParamsBuffer);
 
@@ -115,7 +115,7 @@ namespace sp::vulkan {
     private:
         DeviceContext &device;
         PerfTimer &timer;
-        RenderGraph graph;
+        rg::RenderGraph graph;
 
         void BuildFrameGraph();
         void AddFlatView(ecs::Lock<ecs::Read<ecs::TransformSnapshot, ecs::View>> lock);
@@ -127,7 +127,7 @@ namespace sp::vulkan {
 #endif
 
         void AddScreenshots();
-        RenderGraphResourceID VisualizeBuffer(RenderGraphResourceID sourceID, uint32 arrayLayer = ~0u);
+        rg::ResourceID VisualizeBuffer(rg::ResourceID sourceID, uint32 arrayLayer = ~0u);
 
         void AddSceneState(ecs::Lock<ecs::Read<ecs::Renderable, ecs::TransformSnapshot>> lock);
         void AddGeometryWarp();
@@ -141,8 +141,8 @@ namespace sp::vulkan {
         void AddTonemap();
         void AddMenuOverlay();
 
-        RenderGraphResourceID AddBloom();
-        RenderGraphResourceID AddGaussianBlur(RenderGraphResourceID sourceID,
+        rg::ResourceID AddBloom();
+        rg::ResourceID AddGaussianBlur(rg::ResourceID sourceID,
             glm::ivec2 direction,
             uint32 downsample = 1,
             float scale = 1.0f,
@@ -159,7 +159,7 @@ namespace sp::vulkan {
         struct RenderableGui {
             Tecs::Entity entity;
             shared_ptr<GuiRenderer> renderer;
-            RenderGraphResourceID renderGraphID = ~0u;
+            rg::ResourceID renderGraphID = rg::InvalidResource;
         };
         vector<RenderableGui> guis;
         shared_ptr<GuiRenderer> debugGuiRenderer;
