+++ src/graphics/graphics/vulkan/core/Model.cc
@@ -4,9 +4,9 @@
 #include "core/Logging.hh"
 #include "core/Tracing.hh"
 #include "ecs/EcsImpl.hh"
+#include "graphics/vulkan/VertexLayouts.hh"
 #include "graphics/vulkan/core/CommandContext.hh"
 #include "graphics/vulkan/core/DeviceContext.hh"
-#include "graphics/vulkan/core/Vertex.hh"
 
 #include <tiny_gltf.h>
