+++ src/graphics/graphics/vulkan/render_graph/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(${PROJECT_GRAPHICS_VULKAN_CORE_LIB} PRIVATE
+    PassBuilder.cc
+    RenderGraph.cc
+    Resources.cc
+)
